% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divide_rdname.R
\name{divide}
\alias{divide}
\alias{divide_by}
\alias{divide_by2}
\alias{divide_by3}
\alias{divide_by4}
\title{Scalar Division}
\usage{
divide_by(x, d)

divide_by2(x)

divide_by3(x)

divide_by4(x)
}
\arguments{
\item{x}{The object to be divided}

\item{d}{Scalar to divide \code{x} by.}
}
\value{
The result of the scalar division of \code{x}.
}
\description{
Divide an object by a scalar.
}
\details{
\code{divide_by()} performes generic division by \code{d}.

\code{divide_by2()} performes division by two.

\code{divide_by3()} performes division by three.

\code{divide_by4()} performes division by four.
}
\examples{
x <- matrix(1:6, 3L)

divide_by(x, 1.5)

divide_by2(x)

divide_by3(x)

divide_by4(x)

}
