% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_doc_tpl.R
\name{fun_doc_tpl}
\alias{fun_doc_tpl}
\title{Template Documentation Example}
\usage{
fun_doc_tpl(first_arg, second_arg)
}
\arguments{
\item{first_arg}{Description of the first argument. Also define data type /
class / structure / vectorization expectations where not obvious, and
possibly use (see 'Details').}

\item{second_arg}{Description of the second argument.}
}
\value{
Returns \code{NULL}, invisibly. The function is called for illustration
purposes only.
}
\description{
Show an example of roxygen tags organization, providing documentation best
practices and guidelines, using templates.
}
\details{
More details or context about the function and its behaviour. List
possible side-effects, corner-cases and limitations here. Also use this
section for describing the meaning and usage of arguments when this is too
complex or verbose for the \verb{@param} tags.
}
\section{Custom section}{

The content of a custom section.
}

\examples{
# illustrate through examples how functions can be used
fun_doc_tpl("example", "usage")
}
\references{
Hadley Wickham, Peter Danenberg and Manuel Eugster. roxygen2: In-Line
Documentation for R. \url{https://CRAN.R-project.org/package=roxygen2}.

Hadley Wickham, The tidyverse style guide.
\url{https://style.tidyverse.org/documentation.html}.
}
\seealso{
\code{\link[=fun_doc_tpl]{fun_doc_tpl()}}, \code{\link[roxygen2:roxygenize]{roxygen2::roxygenize()}}. Even if you just put
comma-separated links to functions, don't forget the final period (.).

Other function documentation examples: 
\code{\link{fun_doc}()}
}
\concept{function documentation examples}
