% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/times_describeIn.R
\name{times}
\alias{times}
\alias{times2}
\alias{times3}
\alias{times4}
\title{Scalar Multiplication}
\usage{
times(x, m)

times2(x)

times3(x)

times4(x)
}
\arguments{
\item{x}{The object to be multiplied.}

\item{m}{Scalar to multiply \code{x} by.}
}
\value{
The result of the scalar multiplication of \code{x}.
}
\description{
Multiply an object by a scalar.
}
\section{Functions}{
\itemize{
\item \code{times}: Generic multiplication by \code{m}.

\item \code{times2}: Multiplication by 2.

\item \code{times3}: Multiplication by 3.

\item \code{times4}: Multiplication by 4.
}}

\examples{
x <- matrix(1:6, 3L)

times(x, 1.5)

times2(x)

times3(x)

times4(x)

}
